requirements.txt
asyncio==3.4.3
attrs==23.2.0
autobahn==19.11.2
Automat==22.10.0
bidict==0.22.1
blinker==1.7.0
certifi==2022.12.7
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.7
Flask==3.0.2
Flask-SocketIO==5.3.6
h11==0.14.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
numpy==1.26.4
pandas==2.2.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyjsparser==2.7.1
PyJWT==2.6.0
pyOpenSSL==24.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2024.1
pywin32==306
requests==2.25.1
service-identity==24.1.0
setuptools==69.1.1
simple-websocket==1.0.0
six==1.16.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==1.26.14
websocket-client==1.5.1
websockets==8.1
Werkzeug==3.0.1
wheel==0.42.0
wsproto==1.2.0
zope.interface==6.4.post2
